+++ lib/graphql_rails.rb
@@ -4,7 +4,7 @@ require 'graphql_rails/version'
 require 'graphql_rails/model'
 require 'graphql_rails/router'
 require 'graphql_rails/controller'
-require 'graphql_rails/attribute'
+require 'graphql_rails/attributes'
 
 # wonders starts here
 module GraphqlRails