+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Added/Changed/Deprecated/Removed/Fixed/Security: YOUR CHANGE HERE
+
+## [0.8.0] (2019-09-03)
+
 * Added: permit_input action config with extended list of permitted input options
 * Added: model decorators
 * Added: controller action instrumentation [@povilasjurcys](https://github.com/povilasjurcys)